# Faster collect of sparse vectors by using the Vector constructor